finished core functionality
updated media query of application for devices of width 320px of greater

fixed the reset button active and inactive game states

added a radial background to the winner of a round

may add a 'round-counter' or other animations later on.

the core functionality of the application is finished and it works as it is intended to.